Programmatic Extraction Assessment-Submission/requirements.txt
annotated-types==0.5.0
asttokens==2.2.1
backcall==0.2.0
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
comtypes==1.2.0
confection==0.1.1
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7.post1
decorator==5.1.1
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
fonttools==4.42.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mpmath==1.3.0
murmurhash==1.0.9
nest-asyncio==1.5.7
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
preshed==3.0.8
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
Pygments==2.16.1
PyMuPDF==1.22.5
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.18.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
smart-open==6.3.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.2
sympy==1.12
thinc==8.1.12
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wasabi==1.1.2
wcwidth==0.2.6
